removed a console log, changed my other custom hook to be ts not tsx. moved both my signin and onauth into their own folder so it's easier to find them.
