+++ src/components/Home/Home.tsx
@@ -5,7 +5,7 @@ import styles from './Home.styles';
 import FAQ from './FAQ';
 import HowToPlay from './HowToPlay';
 import CreateRoomButton from './CreateRoomButton';
-import SignInAnonymously from '../Room/SignInAnonymously';
+import SignInAnonymously from '../FirebaseAuth/SignInAnonymously';
 //idk our static folder is so weird and I can't get the image to show
 
 function Home() {
